Feature: CRUDing data to the user's system
References request from #1

Updating types:
    contact id given by the system

This id help us to check if a given contact exits
It helps us also delete the contacts, for those reasons we save It

Some logs are displayed to the console. When the app is running, we can enter debug mode by going to chrome://inspect then inspect our app (REALLY usefull)

We created a ContactAPI containing two functions
    save (to save a Group)
    remove (to effectively remove a group)

To resume things up, the save function loop over contacts to save contacts one by one. If one exists already, delete it before adding it; whereas the remove just loop over to remove the contacts.

It's worth noting that the save function receive an object, so every changes made inside the function will persist after it's call.
Its for this reason I updated the saveGroup function to first save the group to the system that way the _id of each contact will get populated before saving it the the database.